+++ changelog.md
@@ -4,3 +4,9 @@
 - fix export the wrong table
 - fix change combobox_options.json require administrator priviliage
 - fix application.exe do not show the right icon
+- fix save button click do not clear the comboboxes' choices
+- add the following choices / line editor
+  - is gradable
+  - is DR
+  - other diagnosis
+  - confident